// src/CourseGrid/types/CourseSection.ts
export interface CourseSection {
    course_id: string;
    course_code: string;
    grade: number;
    expression: string;
    course_number: string;
    lastfirst: string;
    course_name: string;
    sched_department: string;
    room: string;
    no_of_students: string;
    termid: number;
    teacher: string;
    schoolid: string;
    email: string;
    section_number: string;
    times: string[];
}